Calculation Request, Resources and Controller fixes. Inverters efficiency column fixes

+++ database/migrations/2025_02_19_065023_create_inverter_types_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -14,7 +13,7 @@ return new class extends Migration
         Schema::create('inverter_types', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->integer('efficiency');
+            $table->float('efficiency', 5, 2);
             $table->timestamps();
         });
     }
